fix(3e): resolve compile errors from enum class and deleted copy constructor
- Cast InDelDef to size_t when used as vector index in ErrorStats::AddInDel
- Cast InDelDef to int when used in arithmetic (indel - 2) in Simulator
- Add noexcept move constructor to CoveragePosition to satisfy GCC 15
  vector::resize MoveInsertable requirement (= delete on copy prevents fallback)
- Also explicitly delete copy assignment operator for completeness

+++ reseq/Simulator.cpp
@@ -424,7 +424,7 @@ bool Simulator::FillReadPart(SimRead& sim_read, uintTempSeq template_segment, ui
             }
 
             // Take base call from indel
-            at(sim_read.seq_, par.read_pos_) = indel - 2; // This base call is not stored in par.base_call, because the
+            at(sim_read.seq_, par.read_pos_) = static_cast<int>(indel) - 2; // This base call is not stored in par.base_call, because the
                                                           // normal calls have more predictive power for indel calls
 
             // Update cigar
